Extract workspace render group helpers
